pyproject.toml
[tool.poetry]
name = "resizinator"
version = "0.1.0"
description = ""
authors = ["Vasanth <vasanthn7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pytest = "^8.2.0"
pillow = "^10.3.0"
boto3 = "^1.34.107"
django-storages = "^1.14.3"
redis = "^5.0.4"
celery = "^5.4.0"
gunicorn = "^22.0.0"

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
ruff = "^0.4.4"
pytest-django = "^4.8.0"
django-extensions = "^3.2.3"
moto = "^5.0.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
